feat(lights): RAY-33 #comment Adjust LightSample and implement AmbientLight sampling logic

+++ src/lights/include/ILight.hpp
@@ -13,7 +13,7 @@
 namespace raytracer {
 namespace light {
 struct LightSample {
-    Vector3f wi;
+    Vector3d wi;
     Color radiance;
     double distance = 0.0;
 };
